refactor components into their own files
